upgrade gulp build solution
