+++ gulp/tasks/copy.js
@@ -1,11 +1,10 @@
+/* Copy Gulp task */
 
 "use strict";
 
-const gulp = require('gulp'),
-      changed = require("gulp-changed");
+const gulp = require('gulp');
+const changed = require("gulp-changed");
 
-module.exports = (config)=> [()=>
-    gulp.src(config.src)
-        .pipe(changed(config.dest))
-        .pipe(gulp.dest(config.dest))
-];
+module.exports = (config)=> ()=> gulp.src(config.src)
+    .pipe(changed(config.dest))
+    .pipe(gulp.dest(config.dest));
