--- gulp/tasks/style.js
@@ -1,8 +1,0 @@
-/* Default task for Gulp
- *
- * It just use its settings as an array of tasks names to run
- */
-
-"use strict";
-
-module.exports = (tasks=[]) => [tasks];
