+++ package.json
@@ -15,28 +15,32 @@
     "angular-ui-bootstrap": "^2.5.0",
     "angular-wizard": "^0.10.0",
     "angular1-async-filter": "^1.0.2",
-    "babel-preset-es2015": "^6.18.0",
-    "babelify": "^7.3.0",
     "bootstrap": "^3.3.7",
-    "browserify": "^13.1.1",
     "font-awesome": "^4.7.0",
-    "gulp": "^3.9.1",
+    "ngstorage": "^0.3.11",
+    "rx": "^4.1.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.12.3",
+    "@babel/plugin-transform-runtime": "^7.12.1",
+    "@babel/preset-env": "^7.12.1",
+    "babelify": "^10.0.0",
+    "browserify": "^17.0.0",
+    "gulp": "^4.0.2",
     "gulp-buffer": "0.0.2",
     "gulp-cached": "^1.1.1",
-    "gulp-changed": "^1.3.2",
-    "gulp-connect": "^5.0.0",
-    "gulp-ng-annotate": "^2.0.0",
-    "gulp-replace": "^0.5.4",
-    "gulp-sass": "^3.0.0",
-    "gulp-sourcemaps": "^1.9.1",
-    "gulp-tap": "^0.1.3",
-    "gulp-uglify": "^2.0.0",
-    "gulp-watch": "^4.3.11",
-    "http-proxy-middleware": "^0.17.3",
-    "ngstorage": "^0.3.11",
-    "require-dir": "^0.3.1",
-    "rx": "^4.1.0",
+    "gulp-changed": "^4.0.2",
+    "gulp-connect": "^5.7.0",
+    "gulp-ng-annotate": "^2.1.0",
+    "gulp-replace": "^1.0.0",
+    "gulp-sass": "^4.1.0",
+    "gulp-sourcemaps": "^3.0.0",
+    "gulp-tap": "^2.0.0",
+    "gulp-uglify": "^3.0.2",
+    "http-proxy-middleware": "^1.0.6",
+    "require-dir": "^1.2.0",
     "sass-module-importer": "^1.4.0",
-    "stringify": "^5.1.0"
+    "stringify": "^5.2.0",
+    "undertaker-forward-reference": "^1.0.2"
   }
 }
